+++ kinto/core/authentication.py
@@ -3,20 +3,6 @@ from pyramid import authentication as base_auth
 from kinto.core import utils
 
 
-def prefixed_userid(request):
-    """In Kinto users ids are prefixed with the policy name that is
-    contained in Pyramid Multiauth.
-    If a custom authn policy is used, without authn_type, this method returns
-    the user id without prefix.
-    """
-    # If pyramid_multiauth is used, a ``authn_type`` is set on request
-    # when a policy succesfully authenticates a user.
-    # (see :func:`kinto.core.initialization.setup_authentication`)
-    authn_type = getattr(request, 'authn_type', None)
-    if authn_type is not None:
-        return authn_type + ':' + request.selected_userid
-
-
 class BasicAuthAuthenticationPolicy(base_auth.BasicAuthAuthenticationPolicy):
     """Basic auth implementation.